refactored email verification logic
